+++ benchmarking/blackbox_repository/__init__.py
@@ -1,0 +1,3 @@
+from benchmarking.blackbox_repository.blackbox_offline import BlackboxOffline, deserialize  # noqa: F401
+from benchmarking.blackbox_repository.repository import load, blackbox_list  # noqa: F401
+from benchmarking.blackbox_repository.blackbox_surrogate import add_surrogate  # noqa: F401
