+++ benchmarking/blackbox_repository/conversion_scripts/scripts/icml2020_import.py
@@ -7,8 +7,8 @@ Convert evaluations from
 from typing import Optional
 import pandas as pd
 import numpy as np
-from blackbox_repository.blackbox_offline import serialize, BlackboxOffline
-from blackbox_repository.conversion_scripts.utils import repository_path, upload
+from benchmarking.blackbox_repository.blackbox_offline import serialize, BlackboxOffline
+from benchmarking.blackbox_repository.conversion_scripts.utils import repository_path, upload
 import syne_tune.search_space as sp
