+++ benchmarking/blackbox_repository/conversion_scripts/scripts/nasbench201_import.py
@@ -5,8 +5,8 @@ import pandas as pd
 import numpy as np
 import logging
 
-from blackbox_repository.blackbox_tabular import serialize, BlackboxTabular
-from blackbox_repository.conversion_scripts.utils import repository_path
+from benchmarking.blackbox_repository.blackbox_tabular import serialize, BlackboxTabular
+from benchmarking.blackbox_repository.conversion_scripts.utils import repository_path
 
 from syne_tune import search_space
 from syne_tune.util import catchtime
@@ -224,7 +224,7 @@ def generate_nasbench201(s3_root: Optional[str] = None):
         serialize(bb_dict=bb_dict, path=repository_path / BLACKBOX_NAME)
 
     with catchtime("uploading to S3"):
-        from blackbox_repository.conversion_scripts.utils import upload
+        from benchmarking.blackbox_repository.conversion_scripts import upload
         upload(BLACKBOX_NAME, s3_root=s3_root)
