+++ examples/training_scripts/height_with_cost/train_height_with_cost.py
@@ -21,7 +21,7 @@ import math
 
 from syne_tune.report import Reporter
 from syne_tune.search_space import randint, add_to_argparse
-from benchmarks.checkpoint import resume_from_checkpointed_model, \
+from benchmarking.utils.checkpoint import resume_from_checkpointed_model, \
     checkpoint_model_at_rung_level, add_checkpointing_to_argparse
 from benchmarks.utils import parse_bool
