+++ examples/training_scripts/lstm_wikitext2/lstm_wikitext2.py
@@ -22,7 +22,7 @@ import math
 from syne_tune.report import Reporter
 from syne_tune.search_space import randint, uniform, loguniform, \
     add_to_argparse
-from benchmarks.checkpoint import resume_from_checkpointed_model, \
+from benchmarking.utils.checkpoint import resume_from_checkpointed_model, \
     checkpoint_model_at_rung_level, add_checkpointing_to_argparse, \
     pytorch_load_save_functions
 from benchmarks.utils import parse_bool
