+++ examples/training_scripts/mlp_on_fashion_mnist/mlp_on_fashion_mnist.py
@@ -21,7 +21,7 @@ import time
 from syne_tune.report import Reporter
 from syne_tune.search_space import randint, uniform, loguniform, \
     add_to_argparse
-from benchmarks.checkpoint import resume_from_checkpointed_model, \
+from benchmarking.utils.checkpoint import resume_from_checkpointed_model, \
     checkpoint_model_at_rung_level, add_checkpointing_to_argparse, \
     pytorch_load_save_functions
 from benchmarks.utils import parse_bool
